Loading state for usage table
